+++ sdk/maintenance/azure-mgmt-maintenance/generated_samples/apply_updates_get_parent.py
@@ -42,6 +42,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2023-10-01-preview/ApplyUpdates_GetParent.json
+# x-ms-original-file: 2025-10-01-preview/ApplyUpdates_GetParent.json
 if __name__ == "__main__":
     main()