+++ sdk/maintenance/azure-mgmt-maintenance/generated_samples/configuration_assignments_for_resource_group_delete.py
@@ -37,6 +37,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2023-10-01-preview/ConfigurationAssignmentsForResourceGroup_Delete.json
+# x-ms-original-file: 2025-10-01-preview/ConfigurationAssignmentsForResourceGroup_Delete.json
 if __name__ == "__main__":
     main()