+++ sdk/maintenance/azure-mgmt-maintenance/generated_samples/configuration_assignments_for_subscriptions_update_for_resource.py
@@ -54,6 +54,6 @@ def main():
     print(response)
 
 
-# x-ms-original-file: 2023-10-01-preview/ConfigurationAssignmentsForSubscriptions_UpdateForResource.json
+# x-ms-original-file: 2025-10-01-preview/ConfigurationAssignmentsForSubscriptions_UpdateForResource.json
 if __name__ == "__main__":
     main()